Image_Recognition_Chatbot-master/requirements.txt
torch==2.1.0
torchvision==0.15.0
torchaudio==2.1.0
transformers==4.45.1
pillow==10.4.0
tokenizers==0.20.0
qwen_vl_utils==0.0.8
safetensors==0.4.5
